requirements.txt
streamlit
pandas
numpy
plotly
matplotlib
statsmodels
joblib
pmdarima
mysql-connector-python
sqlalchemy